+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Adjust comfort / container sizes / cost on various pieces.
+
 ## [0.7.2] - 2025-08-05
 
 ### Changed